requirements.txt
Django==2.2.1
django-debug-toolbar==1.11
psycopg2==2.8.2
social-auth-app-django==3.1.0
django-internationalflavor==0.3.1